core/src/main/java/net/cactusthorn/routing/convert/StringConverter.java
package net.cactusthorn.routing.convert;

import java.lang.annotation.Annotation;
import java.lang.reflect.Type;

public final class StringConverter implements Converter {

    @Override //
    public String convert(Class<?> type, Type genericType, Annotation[] annotations, String value) {
        return value;
    }

    @Override //
    public Object convert(Class<?> type, Type genericType, Annotation[] annotations, String[] value) {
        if (value == null) {
            return null;
        }

        String[] array = new String[value.length];
        for (int i = 0; i < value.length; i++) {
            array[i] = convert(type, genericType, annotations, value[i]);
        }
        return array;
    }
}
